refactor : api 타입 적용 - 임시로 any로 설정
